[FC] Remove a deprecation warning.
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/projects/FunctionalCuration@23627 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ src/python/fc/simulations/ranges.py
@@ -116,7 +116,7 @@ class UniformRange(AbstractRange):
         self.step = self.stepExpr.Evaluate(self.env).value
         self.end = self.endExpr.Evaluate(self.env).value
         self.value = self.start
-        self.numberOfOutputs = (round(self.end-self.start)/self.step) + 1
+        self.numberOfOutputs = int((round(self.end-self.start)/self.step)) + 1
 
 
 class VectorRange(AbstractRange):
